requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

cohere==4.51
diskcache==5.6.3
docx2pdf==0.1.8
fastapi==0.110.0
httpx==0.27.0
matplotlib==3.8.3
numpy==1.26.4
openai==1.13.3
pandas==2.2.1
pillow==10.2.0
plotly==5.19.0
pydantic==2.6.3
pytest==8.1.0
python-docx==1.1.0
PyYAML==6.0.1
requests==2.31.0
streamlit==1.32.0
tiktoken==0.6.0
typer==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.27.1
google-auth==2.28.1
python-decouple==3.8
pdfrw==0.4
#opencv-python==4.9.0.80
pillow==10.2.0
seaborn==0.13.2
statsmodels==0.14.1
#python-dotenv


# The following can be used for Okta authentication
# git+https://github.com/sfc-gh-bhess/st_oauth.git
# PyJWT
